Reuse the same search for Budget that in Campaign page
so the result are the same
