+++ plugins/polio/helpers.py
@@ -2,8 +2,11 @@ import json
 from datetime import timedelta
 
 import requests
+from django.db.models import Q
 from django.utils.timezone import now
+from rest_framework import filters
 
+from iaso.models import OrgUnitType, OrgUnit
 from plugins.polio.models import URLCache
 
 
@@ -20,3 +23,23 @@ def get_url_content(url, login, password, minutes=60):
     else:
         j = json.loads(cached_response.content)
     return j
+
+
+class CustomFilterBackend(filters.BaseFilterBackend):
+    def filter_queryset(self, request, queryset, view):
+        search = request.query_params.get("search")
+        if search:
+            country_types = OrgUnitType.objects.countries().only("id")
+            org_units = OrgUnit.objects.filter(
+                name__icontains=search, org_unit_type__in=country_types, path__isnull=False
+            ).only("id")
+
+            query = Q(obr_name__icontains=search) | Q(epid__icontains=search)
+            if len(org_units) > 0:
+                query.add(
+                    Q(initial_org_unit__path__descendants=OrgUnit.objects.query_for_related_org_units(org_units)), Q.OR
+                )
+
+            return queryset.filter(query)
+
+        return queryset
